PIG-3591: Refactor POPackage to separate MR specific code from packaging (mwagner via cheolsoo)
git-svn-id: https://svn.apache.org/repos/asf/pig/trunk@1585283 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/pig/backend/hadoop/executionengine/physicalLayer/relationalOperators/POPartialAgg.java
@@ -229,7 +229,7 @@ public class POPartialAgg extends PhysicalOperator implements Spillable {
                 }
             }
             avgTupleSize = estTotalMem / estTuples;
-            int totalTuples = memLimits.getCacheLimit();
+            long totalTuples = memLimits.getCacheLimit();
             LOG.info("Estimated total tuples to buffer, based on " + estTuples + " tuples that took up " + estTotalMem + " bytes: " + totalTuples);
             firstTierThreshold = (int) (0.5 + totalTuples * (1f - (1f / sizeReduction)));
             secondTierThreshold = (int) (0.5 + totalTuples *  (1f / sizeReduction));
